PAINTROID-757: Research flutter integration_test packages vs flutter test + drive
