+++ test/integration/driver/driver.dart
@@ -1,0 +1,3 @@
+import 'package:integration_test/integration_test_driver_extended.dart';
+
+Future<void> main() => integrationDriver();
